Alpha 3 - E-Prescription/requirements.txt
customtkinter~=5.2.2
mysql-connector-python~=8.3.0
pillow~=10.2.0
reportlab~=4.0.9
qrcode~=7.4.2
passlib~=1.7.4
CTkMessagebox~=2.5